TicTacToe changes
I have updated all of my code to make the program smarter and less hard
coded.  Therefore if any changes are made then the program will still
be able to read the program.  change the winning combos to inituitively
check for who won based on whose turn it is and what symbol is in the
square. I change the draw to be based off of how many turns and if no
winning combos.